+++ setup.py
@@ -32,7 +32,8 @@ with open("README.md", "r", encoding="utf-8") as fh:
     long_description = fh.read()
 
 # Define the minimal classes needed to install and run tigramite
-INSTALL_REQUIRES =  ["numpy==1.21.4", "scipy==1.7.2", "numba==0.53.1", "six"]
+# INSTALL_REQUIRES =  ["numpy==1.21.4", "scipy==1.7.2", "numba==0.53.1", "six"]
+INSTALL_REQUIRES =  ["numpy", "scipy", "numba", "six"]
 # Define all the possible extras needed
 EXTRAS_REQUIRE = {
     "all": [
@@ -59,7 +60,7 @@ CMDCLASS = {"build_ext": UseNumpyHeadersBuildExt}
 setup(
     name="tigramite",
     version="4.3.1.1",
-    packages=["tigramite", "tigramite.independence_tests"],
+    packages=["tigramite", "tigramite.independence_tests", "tigramite.toymodels"],
     license="GNU General Public License v3.0",
     description="Tigramite causal discovery for time series",
     author="Jakob Runge",
